Simplify core facade class (#2743)
* Simplify facade reporting

* Extract binding context creation
