+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/BindingContext.kt
@@ -1,0 +1,50 @@
+package io.gitlab.arturbosch.detekt.core
+
+import org.jetbrains.kotlin.cli.common.messages.AnalyzerWithCompilerReport
+import org.jetbrains.kotlin.cli.common.messages.CompilerMessageLocation
+import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSeverity
+import org.jetbrains.kotlin.cli.common.messages.PlainTextMessageRenderer
+import org.jetbrains.kotlin.cli.common.messages.PrintingMessageCollector
+import org.jetbrains.kotlin.cli.jvm.compiler.KotlinCoreEnvironment
+import org.jetbrains.kotlin.cli.jvm.compiler.NoScopeRecordCliBindingTrace
+import org.jetbrains.kotlin.cli.jvm.compiler.TopDownAnalyzerFacadeForJVM
+import org.jetbrains.kotlin.config.languageVersionSettings
+import org.jetbrains.kotlin.psi.KtFile
+import org.jetbrains.kotlin.resolve.BindingContext
+import org.jetbrains.kotlin.resolve.lazy.declarations.FileBasedDeclarationProviderFactory
+
+internal fun generateBindingContext(
+    environment: KotlinCoreEnvironment,
+    classpath: List<String>,
+    files: List<KtFile>
+): BindingContext {
+    if (classpath.isEmpty()) {
+        return BindingContext.EMPTY
+    }
+
+    val analyzer = AnalyzerWithCompilerReport(
+        PrintingMessageCollector(System.err, DetektMessageRenderer, true),
+        environment.configuration.languageVersionSettings
+    )
+    analyzer.analyzeAndReport(files) {
+        TopDownAnalyzerFacadeForJVM.analyzeFilesWithJavaIntegration(
+            environment.project, files, NoScopeRecordCliBindingTrace(),
+            environment.configuration, environment::createPackagePartProvider,
+            ::FileBasedDeclarationProviderFactory
+        )
+    }
+    return analyzer.analysisResult.bindingContext
+}
+
+private object DetektMessageRenderer : PlainTextMessageRenderer() {
+    override fun getName() = "detekt message renderer"
+    override fun getPath(location: CompilerMessageLocation) = location.path
+    override fun render(
+        severity: CompilerMessageSeverity,
+        message: String,
+        location: CompilerMessageLocation?
+    ): String {
+        if (!severity.isError) return ""
+        return super.render(severity, message, location)
+    }
+}
